# Week7/backend/openapi_server/models/book.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Book(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, title=None, author=None, available=None):  # noqa: E501
        """Book - a model defined in OpenAPI

        :param id: The id of this Book.  # noqa: E501
        :type id: int
        :param title: The title of this Book.  # noqa: E501
        :type title: str
        :param author: The author of this Book.  # noqa: E501
        :type author: str
        :param available: The available of this Book.  # noqa: E501
        :type available: bool
        """
        self.openapi_types = {
            'id': int,
            'title': str,
            'author': str,
            'available': bool
        }

        self.attribute_map = {
            'id': 'id',
            'title': 'title',
            'author': 'author',
            'available': 'available'
        }

        self._id = id
        self._title = title
        self._author = author
        self._available = available

    @classmethod
    def from_dict(cls, dikt) -> 'Book':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Book of this Book.  # noqa: E501
        :rtype: Book
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Book.


        :return: The id of this Book.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Book.


        :param id: The id of this Book.
        :type id: int
        """

        self._id = id

    @property
    def title(self) -> str:
        """Gets the title of this Book.


        :return: The title of this Book.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this Book.


        :param title: The title of this Book.
        :type title: str
        """

        self._title = title

    @property
    def author(self) -> str:
        """Gets the author of this Book.


        :return: The author of this Book.
        :rtype: str
        """
        return self._author

    @author.setter
    def author(self, author: str):
        """Sets the author of this Book.


        :param author: The author of this Book.
        :type author: str
        """

        self._author = author

    @property
    def available(self) -> bool:
        """Gets the available of this Book.


        :return: The available of this Book.
        :rtype: bool
        """
        return self._available

    @available.setter
    def available(self, available: bool):
        """Sets the available of this Book.


        :param available: The available of this Book.
        :type available: bool
        """

        self._available = available