# Week7/backend/openapi_server/models/book_borrowed.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class BookBorrowed(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, member_id=None, book_id=None, borrow_date=None, return_date=None):  # noqa: E501
        """BookBorrowed - a model defined in OpenAPI

        :param id: The id of this BookBorrowed.  # noqa: E501
        :type id: int
        :param member_id: The member_id of this BookBorrowed.  # noqa: E501
        :type member_id: int
        :param book_id: The book_id of this BookBorrowed.  # noqa: E501
        :type book_id: int
        :param borrow_date: The borrow_date of this BookBorrowed.  # noqa: E501
        :type borrow_date: str
        :param return_date: The return_date of this BookBorrowed.  # noqa: E501
        :type return_date: str
        """
        self.openapi_types = {
            'id': int,
            'member_id': int,
            'book_id': int,
            'borrow_date': str,
            'return_date': str
        }

        self.attribute_map = {
            'id': 'id',
            'member_id': 'member_id',
            'book_id': 'book_id',
            'borrow_date': 'borrow_date',
            'return_date': 'return_date'
        }

        self._id = id
        self._member_id = member_id
        self._book_id = book_id
        self._borrow_date = borrow_date
        self._return_date = return_date

    @classmethod
    def from_dict(cls, dikt) -> 'BookBorrowed':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BookBorrowed of this BookBorrowed.  # noqa: E501
        :rtype: BookBorrowed
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this BookBorrowed.


        :return: The id of this BookBorrowed.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this BookBorrowed.


        :param id: The id of this BookBorrowed.
        :type id: int
        """

        self._id = id

    @property
    def member_id(self) -> int:
        """Gets the member_id of this BookBorrowed.


        :return: The member_id of this BookBorrowed.
        :rtype: int
        """
        return self._member_id

    @member_id.setter
    def member_id(self, member_id: int):
        """Sets the member_id of this BookBorrowed.


        :param member_id: The member_id of this BookBorrowed.
        :type member_id: int
        """

        self._member_id = member_id

    @property
    def book_id(self) -> int:
        """Gets the book_id of this BookBorrowed.


        :return: The book_id of this BookBorrowed.
        :rtype: int
        """
        return self._book_id

    @book_id.setter
    def book_id(self, book_id: int):
        """Sets the book_id of this BookBorrowed.


        :param book_id: The book_id of this BookBorrowed.
        :type book_id: int
        """

        self._book_id = book_id

    @property
    def borrow_date(self) -> str:
        """Gets the borrow_date of this BookBorrowed.


        :return: The borrow_date of this BookBorrowed.
        :rtype: str
        """
        return self._borrow_date

    @borrow_date.setter
    def borrow_date(self, borrow_date: str):
        """Sets the borrow_date of this BookBorrowed.


        :param borrow_date: The borrow_date of this BookBorrowed.
        :type borrow_date: str
        """

        self._borrow_date = borrow_date

    @property
    def return_date(self) -> str:
        """Gets the return_date of this BookBorrowed.


        :return: The return_date of this BookBorrowed.
        :rtype: str
        """
        return self._return_date

    @return_date.setter
    def return_date(self, return_date: str):
        """Sets the return_date of this BookBorrowed.


        :param return_date: The return_date of this BookBorrowed.
        :type return_date: str
        """

        self._return_date = return_date